Refactoring and commenting
